AArch64: Use negated constant value for add/sub node if it is more concise
This commit changes `genericBinaryEvaluator` to use negated value for add/sub
operation if the negated value can be loaded into the register with the fewer
instructions than original value.
Also, `genericBinaryEvaluator` to set a register to the constant node
if the constant cannot be encoded into the instruction and the negated value is not used.
This change prevents instructions to load the constant into the register from
generated multiple times even if the constant node has multiple reference counts.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>